adding command 'add'

+++ data/expense.go
@@ -3,8 +3,8 @@ package data
 // Package data provides structures for handling financial data.
 
 type Expense struct {
-	ID          int     `json:"id"`
-	Description string  `json:"description"`
-	Amount      float64 `json:"amount"`
-	Date        string  `json:"date"`
+	ID          int    `json:"id"`
+	Description string `json:"description"`
+	Amount      int    `json:"amount"`
+	Date        string `json:"date"`
 }
